--- src/main/java/com/tinkoff/edu/app/ILoanCalcService.java
@@ -1,5 +1,0 @@
-package com.tinkoff.edu.app;
-
-public interface ILoanCalcService {
-    LoanModel createRequest(LoanRequest loanRequest);
-}
