+++ src/main/java/com/tinkoff/edu/app/loan/repository/ILoanCalcRepository.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.loan.repository;
 
 public interface ILoanCalcRepository {
     int save();
