+++ src/main/java/com/tinkoff/edu/app/loan/repository/impl/LoanCalcRepository.java
@@ -1,4 +1,6 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.loan.repository.impl;
+
+import com.tinkoff.edu.app.loan.repository.ILoanCalcRepository;
 
 public class LoanCalcRepository implements ILoanCalcRepository {
     private int requestId;
