+++ src/main/java/com/tinkoff/edu/app/loan/types/LoanType.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.loan.types;
 
 public enum LoanType {
     PERSON, IP, OOO
