+++ src/main/java/com/tinkoff/edu/app/loan/types/ResponseType.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.loan.types;
 
 public enum ResponseType {
     APPROVED, DENIED
